+++ main.go
@@ -22,19 +22,16 @@ THE SOFTWARE.
 package main
 
 import (
+	"log/slog"
 	"os"
 
 	"github.com/gnames/gnverifier/cmd"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
+	"github.com/lmittmann/tint"
 )
 
 func main() {
-	log.Logger = log.Output(
-		zerolog.ConsoleWriter{
-			Out:        os.Stderr,
-			TimeFormat: "15:04:05",
-		},
-	)
+	handle := slog.New(tint.NewHandler(os.Stderr, nil))
+	slog.SetDefault(handle)
+
 	cmd.Execute()
 }