+++ pkg/ent/output/output_test.go
@@ -54,7 +54,7 @@ func TestOutput(t *testing.T) {
 				assert.Contains(t, res, "bestResult", "pretty 2")
 				assert.Contains(t, res, "results", "pretty 3")
 			},
-			linesNum: 116,
+			linesNum: 119,
 		},
 		{
 			msg:    "compact",